+++ src/QuickInfo/RazorQuickInfoSource.cs
@@ -2,10 +2,11 @@
 using Microsoft.VisualStudio.Text;
 using System;
 using System.Linq;
+using TailwindCSSIntellisense.Completions;
 
 namespace TailwindCSSIntellisense.QuickInfo;
 
-internal class RazorQuickInfoSource(ITextBuffer textBuffer, DescriptionGenerator descriptionGenerator) : QuickInfoSource(textBuffer, descriptionGenerator)
+internal class RazorQuickInfoSource(ITextBuffer textBuffer, DescriptionGenerator descriptionGenerator, CompletionUtilities completionUtilities) : QuickInfoSource(textBuffer, descriptionGenerator, completionUtilities)
 {
     protected override bool IsInClassScope(IAsyncQuickInfoSession session, out SnapshotSpan? span)
     {