+++ src/QuickInfo/RazorQuickInfoSourceProvider.cs
@@ -2,6 +2,7 @@
 using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Utilities;
 using System.ComponentModel.Composition;
+using TailwindCSSIntellisense.Completions;
 
 namespace TailwindCSSIntellisense.QuickInfo;
 
@@ -16,8 +17,11 @@ internal sealed class RazorQuickInfoSourceProvider : IAsyncQuickInfoSourceProvid
     [Import]
     public DescriptionGenerator DescriptionGenerator { get; set; }
 
+    [Import]
+    public CompletionUtilities CompletionUtilities { get; set; }
+
     public IAsyncQuickInfoSource TryCreateQuickInfoSource(ITextBuffer textBuffer)
     {
-        return textBuffer.Properties.GetOrCreateSingletonProperty(() => new RazorQuickInfoSource(textBuffer, DescriptionGenerator));
+        return textBuffer.Properties.GetOrCreateSingletonProperty(() => new RazorQuickInfoSource(textBuffer, DescriptionGenerator, CompletionUtilities));
     }
 }